lambda/s3-triggered-lambda/requirements.txt
boto3==1.28.57
Pillow==9.5.0
piexif==1.1.3
requests==2.31.0